refactor common UI code
