+++ qiskit_aqua_interfaces/aqua/user_interface/_aquathread.py
@@ -23,7 +23,7 @@ import platform
 import os
 import subprocess
 import psutil
-from qiskit_aqua_interfaces.aqua.user_interface import GUIProvider
+from qiskit_aqua_interfaces.user_interface import GUIProvider
 
 logger = logging.getLogger(__name__)
