+++ qiskit_aqua_interfaces/chemistry/user_interface/_controller.py
@@ -16,7 +16,7 @@
 
 from tkinter import messagebox
 import logging
-from qiskit_aqua_interfaces.aqua.user_interface import BaseController
+from qiskit_aqua_interfaces.user_interface import BaseController
 from ._model import Model
 
 logger = logging.getLogger(__name__)
