+++ qiskit_aqua_interfaces/chemistry/user_interface/command_line.py
@@ -18,7 +18,7 @@ import sys
 import logging
 import tkinter as tk
 from tkinter import messagebox
-from qiskit_aqua_interfaces.aqua.user_interface import MainView
+from qiskit_aqua_interfaces.user_interface import MainView
 from qiskit_aqua_interfaces._extras_require import _check_extra_requires
