+++ qiskit_aqua_interfaces/user_interface/__init__.py
@@ -2,7 +2,7 @@
 
 # This code is part of Qiskit.
 #
-# (C) Copyright IBM 2018, 2019.
+# (C) Copyright IBM 2019.
 #
 # This code is licensed under the Apache License, Version 2.0. You may
 # obtain a copy of this license in the LICENSE.txt file in the root directory
@@ -12,8 +12,18 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
-"""Qiskit Chemistry User Interface imports"""
+"""Common User Interface"""
 
-from ._uipreferences import UIPreferences
+from .guiprovider import GUIProvider
+from .base_controller import BaseController
+from .base_model import BaseModel
+from ._customwidgets import TextPopup
+from ._sectionpropertiesview import SectionPropertiesView
+from ._mainview import MainView
 
-__all__ = ['UIPreferences']
+__all__ = ['GUIProvider',
+           'BaseController',
+           'BaseModel',
+           'TextPopup',
+           'SectionPropertiesView',
+           'MainView']
