+++ qiskit_aqua_interfaces/user_interface/_credentialsview.py
