+++ qiskit_aqua_interfaces/user_interface/_customwidgets.py
