+++ qiskit_aqua_interfaces/user_interface/_mainview.py
@@ -12,6 +12,8 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
+"""Root main view"""
+
 import sys
 import tkinter as tk
 import tkinter.messagebox as tkmb
