+++ qiskit_aqua_interfaces/user_interface/_sectionpropertiesview.py
@@ -12,6 +12,8 @@
 # copyright notice, and modified files need to carry a notice indicating
 # that they have been altered from the originals.
 
+"""Section Properties View"""
+
 import tkinter as tk
 import tkinter.ttk as ttk
 from ._toolbarview import ToolbarView
