+++ qiskit_aqua_interfaces/user_interface/_sectionsview.py
