+++ qiskit_aqua_interfaces/user_interface/guiprovider.py
